feat(import-export): EPSG not mandatory and fix encoding issue (#428)
