+++ packages/utils/src/lib/index.ts
@@ -2,6 +2,7 @@ export * from './base64';
 export * from './clipboard';
 export * from './change';
 export * from './change.interface';
+export * from './file';
 export * from './object-utils';
 export * from './strenum';
 export * from './string-utils';
